Move the signal-analysis helpers out of the shipped package
correlation_peak and spectral_cosine are consumed only by tests, so
they move from reachy_mini.media.audio_utils to tests/audio_helpers.py
(importable from both test trees via the root conftest.py's sys.path
entry). Side effect worth having: the hardware test now imports nothing
that isn't in a stock reachy-mini install, so it runs against an
unmodified robot.

load_audio_mono stays in the package: it is the read counterpart of
save_audio_to_wav (which stdlib wave cannot read back, IEEE-float WAV).
